File info from HostClient to CentralServer
The HostClient now sends both user info, which is used to update the CentralServer's user table textfile if the hostname is new, and a list of files and their descriptions that the HostClient wants to share. The next step will be to have the central server make a similar textfile table for storing file info, and then a keyword search which will pull out relevant results based on the keywords sent by HostClient, and return those results to the HostClient for the table. 
The "sharedFiles.txt" file contains the file descriptions sent to the CentralServer.



To compile the central server:
javac -d . User.Java
javac -d . SharedFile.java
javac -d . CentralServer.java